fixing space saving calculation

+++ asgn6/decode.c
@@ -100,8 +100,8 @@ int main(int argc, char **argv) {
     if (verbose == true) {
         fprintf(stderr, "compressed file size = %lu bytes\n", infile_stat.st_size);
         fprintf(stderr, "decompressed file size = %lu bytes\n", outfile_stat.st_size);
-        uint64_t saving = 100 * (1 - (infile_stat.st_size / outfile_stat.st_size));
-        fprintf(stderr, "space saving = %" PRIu64 "%%\n", saving);
+        double saving = 100 * (1 - ((double)infile_stat.st_size / (double)outfile_stat.st_size));
+        fprintf(stderr, "space saving = %.2f%%\n", saving);
     }
 
     delete_tree(&root);
